copy trunk to release branch in preparation for repurposing trunk.
in the past, we’ve been using trunk more or less as “that place with the
100% stable, production-ready stuff,” which really is the definition of a
release branch. copying it to a release branch.

since this was never put through any formal release process, it doesn’t have
a release number. calling it release-0.support.dataprep in the hopes that
this name will make it clear that:
 a) it’s a release branch,
 b) it’s chronologically before any 1.x release branch, and
 c) its purpose is to support the dataprep process.

once we’ve modified all the current production workflow to use this branch
instead of trunk, we’ll repurpose trunk.

note that this release branch may become unnecessary when we finish and
release the dataprep milestone of the finding aids project.