copy irish to a new name closer to release branch naming conventions, in preparation for deleing old irish branch.
we’ve been treating irish like a release branch for a while now. this new
name comes closer to following the conventions of a release branch. of
course, since it never went through a formal release process, it doesn’t
have a release number, so some variance from the conventional naming was
necessary.

once all references to the old “irish” release-branch-like-directory are
gone, we’ll delete it. hopefully that’ll be very soon.